Move files and fix naming
