+++ tests/allcalrc
