+++ tinycal/__init__.py
@@ -3,4 +3,4 @@ Package description and configuration.
 """
 
 __name__ = 'tinycal'
-CALRC = ('~/.config/calrc', '~/.calrc')
+CALRCS = ('~/.config/calrc', '~/.calrc')
